senate-bot/backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
python-multipart
mysql-connector-python
